requirements.txt
PyMuPDF==1.18.2
Pillow==8.3.1
tqdm==4.47.0